+++ FoxProject/lib/jxcapture-samples.jar
