+++ FoxProject/nbproject/genfiles.properties
@@ -1,0 +1,8 @@
+build.xml.data.CRC32=20e15e79
+build.xml.script.CRC32=75ec73f3
+build.xml.stylesheet.CRC32=8064a381@1.79.1.48
+# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
+# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
+nbproject/build-impl.xml.data.CRC32=20e15e79
+nbproject/build-impl.xml.script.CRC32=e966767c
+nbproject/build-impl.xml.stylesheet.CRC32=05530350@1.79.1.48
