+++ FoxProject/nbproject/private/private.properties
@@ -1,0 +1,2 @@
+compile.on.save=true
+user.properties.file=/home/rootg/.netbeans/8.1/build.properties
