+++ FoxProject/nbproject/project.properties
@@ -1,0 +1,86 @@
+annotation.processing.enabled=true
+annotation.processing.enabled.in.editor=false
+annotation.processing.processor.options=
+annotation.processing.processors.list=
+annotation.processing.run.all.processors=true
+annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
+build.classes.dir=${build.dir}/classes
+build.classes.excludes=**/*.java,**/*.form
+# This directory is removed when the project is cleaned:
+build.dir=build
+build.generated.dir=${build.dir}/generated
+build.generated.sources.dir=${build.dir}/generated-sources
+# Only compile against the classpath explicitly listed here:
+build.sysclasspath=ignore
+build.test.classes.dir=${build.dir}/test/classes
+build.test.results.dir=${build.dir}/test/results
+# Uncomment to specify the preferred debugger connection transport:
+#debug.transport=dt_socket
+debug.classpath=\
+    ${run.classpath}
+debug.test.classpath=\
+    ${run.test.classpath}
+# Files in build.classes.dir which should be excluded from distribution jar
+dist.archive.excludes=
+# This directory is removed when the project is cleaned:
+dist.dir=dist
+dist.jar=${dist.dir}/FoxProject.jar
+dist.javadoc.dir=${dist.dir}/javadoc
+excludes=
+file.reference.jniwrap-3.11.jar=lib/jniwrap-3.11.jar
+file.reference.jxcapture-3.3-javadoc.zip=lib/jxcapture-3.3-javadoc.zip
+file.reference.jxcapture-3.3.jar=lib/jxcapture-3.3.jar
+file.reference.jxcapturedemo.jar=lib/jxcapturedemo.jar
+file.reference.slf4j-api-1.5.8.jar=lib/slf4j-api-1.5.8.jar
+file.reference.slf4j-simple-1.5.8.jar=lib/slf4j-simple-1.5.8.jar
+includes=**
+jar.compress=false
+javac.classpath=\
+    ${file.reference.jniwrap-3.11.jar}:\
+    ${file.reference.jxcapture-3.3-javadoc.zip}:\
+    ${file.reference.slf4j-api-1.5.8.jar}:\
+    ${file.reference.slf4j-simple-1.5.8.jar}:\
+    ${file.reference.jxcapture-3.3.jar}:\
+    ${file.reference.jxcapturedemo.jar}
+# Space-separated list of extra javac options
+javac.compilerargs=
+javac.deprecation=false
+javac.external.vm=true
+javac.processorpath=\
+    ${javac.classpath}
+javac.source=1.8
+javac.target=1.8
+javac.test.classpath=\
+    ${javac.classpath}:\
+    ${build.classes.dir}
+javac.test.processorpath=\
+    ${javac.test.classpath}
+javadoc.additionalparam=
+javadoc.author=false
+javadoc.encoding=${source.encoding}
+javadoc.noindex=false
+javadoc.nonavbar=false
+javadoc.notree=false
+javadoc.private=false
+javadoc.splitindex=true
+javadoc.use=true
+javadoc.version=false
+javadoc.windowtitle=
+main.class=GirisEkrani
+manifest.file=manifest.mf
+meta.inf.dir=${src.dir}/META-INF
+mkdist.disabled=false
+platform.active=default_platform
+run.classpath=\
+    ${javac.classpath}:\
+    ${build.classes.dir}
+# Space-separated list of JVM arguments used when running the project.
+# You may also define separate properties like run-sys-prop.name=value instead of -Dname=value.
+# To set system properties for unit tests define test-sys-prop.name=value:
+run.jvmargs=
+run.test.classpath=\
+    ${javac.test.classpath}:\
+    ${build.test.classes.dir}
+source.encoding=UTF-8
+src.dir=src
+test.src.dir=test
